Generate .RAW image tiles of a heightmap

This function combines the main functions in the [terrainr] package to create .RAW image tiles of a USGS digital elevation model (DEM) covering an area of interest.

generate_raw_tiles(
  data = NULL,
  lat = NULL,
  lng = NULL,
  raw_output_prefix,
  bbox = NULL,
  buffer_distance = NULL,
  distance_unit = "meters",
  unity_friendly = TRUE,
  tif_filename = NULL,
  save_tif = TRUE
)

Arguments

data
Optionally, a dataframe containing vectors of latitude and longitude.

lat
If data is not NULL, the name of the column containing latitude values. If data is NULL, a vector of latitude values.

lng
If data is not NULL, the name of the column containing longitude values. If data is NULL, a vector of longitude values.

raw_output_prefix
A string indicating the prefix to label output files with. Files are named in the pattern output_prefix_x_y.raw where x and y are the x and y position in pixels of the upper left corner.

bbox
Optionally, a bounding box representing the area to pull a DEM for. If provided, arguments to data, lat, and lng are ignored.

buffer_distance
Optionally, a numeric value indicating how much to pad the bounding box by.

distance_unit
A string passed to [terrainr::convert_distance] specifying the units of the buffer_distance argument.

unity_friendly
Logical: should the bounding box be padded to ensure the output .raw tiles can be easily imported to Unity?

tif_filename
The path to save the DEM raster file to. Must end with .tif or .tiff. Ignored if save_tif is FALSE.

save_tif
Logical: should the DEM raster be saved as a TIF in addition to .raw tiles?