Get bounding box for set of coordinate points

Get bounding box for set of coordinate points

get_coord_bbox(data = NULL, lat, lng)

Arguments

data
Optionally, a dataframe containing vectors of latitude and longitude.

lat
If data is not NULL, the name of the column containing latitude values. If data is NULL, a vector of latitude values.

lng
If data is not NULL, the name of the column containing longitude values. If data is NULL, a vector of longitude values.

Value

A list of length 2, containing the bottom-left (named "bl") and top-right (named "tr") coordinates of the bounding box.

Examples

df <- data.frame(
  lat = c(44.05771, 44.18475),
  lng = c(-73.99212, -73.81515)
)
get_coord_bbox(df, lat, lng)

#> An object of class "terrainr_bounding_box"
#> Slot "bl":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.05771 
#> 
#> Slot "lng":
#>       lng 
#> -73.99212 
#> 
#> 
#> Slot "tr":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.18475 
#> 
#> Slot "lng":
#>       lng 
#> -73.81515 
#> 
#> 

get_coord_bbox(lat = df$lat, lng = df$lng)

#> An object of class "terrainr_bounding_box"
#> Slot "bl":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.05771 
#> 
#> Slot "lng":
#>       lng 
#> -73.99212 
#> 
#> 
#> Slot "tr":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.18475 
#> 
#> Slot "lng":
#>       lng 
#> -73.81515 
#> 
#>