Merge multiple raster files into a single raster

Some functions like [get_tiles] return multiple separate files when it can be useful to have a single larger raster instead. This function is a thin wrapper over [sf::gdal_utils(util = "warp")], making it easy to collapse those multiple raster files into a single TIFF.

merge_rasters(
  input_rasters,
  output_raster = tempfile(fileext = ".tif"),
  options = character(0)
)

Arguments

input_rasters
A character vector containing the file paths to the georeferenced rasters you want to use.

output_raster
The file path to save the merged georeferenced raster to.

options
Optionally, a character vector of options to be passed directly to [sf::gdal_utils].

Value

`output_raster`, invisibly.

See also

Examples

if (FALSE) {
simulated_data <- data.frame(
lat = c(44.10379, 44.17573),
lng = c(-74.01177, -73.91171)
)

simulated_data <- sf::st_as_sf(simulated_data, coords = c("lng", "lat"))

img_files <- get_tiles(simulated_data)
merge_rasters(img_files[[1]])

}