Convert radians to degrees

Convert radians to degrees

rad_to_deg(rad)

Arguments

rad
A vector of values, in radians, to convert to decimal degrees

Value

A vector of the same length in decimal degrees

See also

Examples

rad_to_deg(2 * base::pi)

#> [1] 360

rad_to_deg(deg_to_rad(360))

#> [1] 360