Add a uniform buffer around a bounding box

This function calculates the great circle distance both corners of your bounding box are from the centroid and extends those by a set distance. Due to using Haversine "great circle" distance, calculations will not be exact.

add_bbox_buffer(bbox, distance, distance_unit = "meters", divisible = NULL)

Arguments

bbox
The original bounding box to add a buffer around. If not already a terrainr_bounding_box object, will be converted.

distance
The distance to add to the buffer.

distance_unit
The units of the distance to add to the buffer, passed to convert_distance.

divisible
Numeric: Extend the top right and bottom left corners so that each side is divisible by divisible. Leave set to NULL to not extend. This argument is in the same units as distance_unit.

Value

A terrainr_bounding_box object.

See also

Examples

add_bbox_buffer(
  list(
    c(lat = 44.04905, lng = -74.01188),
    c(lat = 44.17609, lng = -73.83493)
  ),
  10
)

#> An object of class "terrainr_bounding_box"
#> Slot "bl":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.04899 
#> 
#> Slot "lng":
#>       lng 
#> -74.01197 
#> 
#> 
#> Slot "tr":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.17615 
#> 
#> Slot "lng":
#>       lng 
#> -73.83484 
#> 
#>