Get bounding box for set of coordinate points.

This function returns a terrainr_bounding_box object representing the bottom left and upper right corners of the smallest rectangle containing your data. If you only have one data point for either latitude or longitude, this function will buffer it in both directions by 1e-10 in order to return a rectangle with a real "bottom left" and "upper right".

get_coord_bbox(data = NULL, lat, lng, na.rm = NULL)

Arguments

data
Optionally, a dataframe containing vectors of latitude and longitude.

lat
If data is not NULL, the name of the column containing latitude values. If data is NULL, a vector of latitude values.

lng
If data is not NULL, the name of the column containing longitude values. If data is NULL, a vector of longitude values.

na.rm
Logical: Silently remove NA values? If NULL, the default, will warn if there are NAs. If FALSE, will raise an error on NA.

Value

A terrainr_bounding_box object.

See also

Examples

df <- data.frame(
  lat = c(44.05771, 44.18475),
  lng = c(-73.99212, -73.81515)
)
get_coord_bbox(df, "lat", "lng")

#> An object of class "terrainr_bounding_box"
#> Slot "bl":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.05771 
#> 
#> Slot "lng":
#>       lng 
#> -73.99212 
#> 
#> 
#> Slot "tr":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.18475 
#> 
#> Slot "lng":
#>       lng 
#> -73.81515 
#> 
#> 

get_coord_bbox(lat = df$lat, lng = df$lng)

#> An object of class "terrainr_bounding_box"
#> Slot "bl":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.05771 
#> 
#> Slot "lng":
#>       lng 
#> -73.99212 
#> 
#> 
#> Slot "tr":
#> An object of class "terrainr_coordinate_pair"
#> Slot "lat":
#>      lat 
#> 44.18475 
#> 
#> Slot "lng":
#>       lng 
#> -73.81515 
#> 
#>