Merge multiple raster files into a single raster

Some functions like get_tiles return multiple separate files when it can be useful to have a single larger raster instead. This function will collapse those multiple raster files into a single TIFF.

merge_rasters(
  input_rasters,
  output_raster = NULL,
  input_images = NULL,
  output_image = NULL,
  overwrite = TRUE,
  merge_raster = TRUE
)

Arguments

input_rasters
A character vector containing the file paths to the georeferenced rasters you want to use.

output_raster
The file path to save the merged georeferenced raster to. Must be a TIFF file. Ignored if merge_raster is not TRUE.

input_images
A character vector, the same length as input_rasters, containing the file paths to non-georeferenced images to merge. It's assumed that each image in the vector corresponds to the raster in the same position in input_rasters, and these images will be assigned the same extent and CRS. If you don't want to merge any images, leave as NULL.

output_image
The file path to save the merged images to. Must be a TIFF file. Leave as NULL if you aren't merging images.

overwrite
Logical: overwrite the output files if they exist?

merge_raster
Logical: Should the files provided to input_rasters be merged? Set to FALSE if you're only providing these input tiles to georeference tiles in input_images.

Value

A named list containing the file paths outputs were written to.

Details

Additionally, some outputs from get_tiles (such as when service = "ortho") are not georeferenced, making future processing with them harder. This function can georeference those images using a reference raster, then merge them into a single object for further processing.

See also

Examples

if (FALSE) {
tile_1 <- terrainr_bounding_box(
  bl = c(lat = 44.10379, lng = -74.01177),
  tr = c(lat = 44.17573, lng = -73.91171)
)

tile_2 <- terrainr_bounding_box(
  bl = c(lat = 44.03184, lng = -74.01177),
  tr = c(lat = 44.10379, lng = -73.91171)
)
raster_files <- c(tempfile(), tempfile())
img_bin <- lapply(
  c(tile_1, tile_2),
  function(x) hit_national_map_api(x, 8000, 8000, "3DEPElevation")
)
writeBin(img_bin[[1]], raster_files[[1]])
writeBin(img_bin[[2]], raster_files[[2]])
merge_rasters(raster_files, tempfile())
}