Crop a raster and convert the output tiles into new formats.

This function crops input raster files into smaller square tiles and then converts them into either .png or .raw files which are ready to be imported into the Unity game engine.

raster_to_raw_tiles(input_file, output_prefix, side_length = 4097, raw = TRUE)

Arguments

input_file
File path to the input TIFF file to convert.

output_prefix
The file path to prefix output tiles with.

side_length
The side length, in pixels, for the .raw tiles.

raw
Logical: Convert the cropped tiles to .raw? When FALSE returns a .png.

Value

Invisibly, a character vector containing the file paths that were written to.

See also

Examples

if (FALSE) {
simulated_data <- data.frame(
  id = seq(1, 100, 1),
  lat = runif(100, 44.04905, 44.17609),
  lng = runif(100, -74.01188, -73.83493)
)
bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
bbox <- add_bbox_buffer(bbox, 100)
output_files <- get_tiles(bbox)
temptiff <- tempfile(fileext = ".tif")
merge_rasters(output_files["3DEPElevation"][[1]], temptiff)
raster_to_raw_tiles(temptiff, tempfile())
}